+++ src/tracks/IsoformAndVariantTrack.ts
@@ -11,6 +11,7 @@ import {
   renderTrackDescription,
 } from '../services/TrackService'
 import {
+  filterVariantData,
   generateDelinsPoint,
   generateInsertionPoint,
   generateSnvPoints,
@@ -24,19 +25,19 @@ import {
   renderVariantDescriptions,
 } from '../services/VariantService'
 
+import { drawIsoforms } from './trackUtils'
+import BaseTrack from './BaseTrack'
+
 import type { VariantFeature } from '../services/VariantService'
 import type { SimpleFeatureSerialized } from '../services/types'
 import type { Selection } from 'd3'
 
-export default class IsoformAndVariantTrack {
+export default class IsoformAndVariantTrack extends BaseTrack {
   private trackData: SimpleFeatureSerialized[]
   private variantData: VariantFeature[]
-  private viewer: Selection<SVGGElement, unknown, HTMLElement | null, undefined>
-  private width: number
   private variantFilter: string[]
   private isoformFilter: string[]
   private initialHighlight?: string[]
-  private height: number
   private transcriptTypes: string[]
   private variantTypes: string[]
   private binRatio: number
@@ -69,14 +70,12 @@ export default class IsoformAndVariantTrack {
     trackData?: SimpleFeatureSerialized[]
     variantData?: VariantFeature[]
   }) {
+    super({ viewer, width, height })
     this.trackData = trackData ?? []
     this.variantData = variantData ?? []
-    this.viewer = viewer
-    this.width = width
     this.variantFilter = variantFilter
     this.isoformFilter = isoformFilter
     this.initialHighlight = initialHighlight
-    this.height = height
     this.transcriptTypes = transcriptTypes
     this.variantTypes = variantTypes
     this.binRatio = binRatio
@@ -87,7 +86,7 @@ export default class IsoformAndVariantTrack {
     const isoformFilter = this.isoformFilter
     let isoformData = this.trackData
     const initialHighlight = this.initialHighlight
-    const variantData = this.filterVariantData(
+    const variantData = filterVariantData(
       this.variantData,
       this.variantFilter,
     )
@@ -499,385 +498,31 @@ export default class IsoformAndVariantTrack {
       .attr('transform', `translate(0,${newTrackPosition})`)
       .attr('class', 'track')
 
-    let row_count = 0
-    const used_space = [] as string[][]
-    let fmin_display = -1
-    let fmax_display = -1
+    const returnedHeight = drawIsoforms({
+      track,
+      isoformData,
+      x,
+      width,
+      isoformFilter,
+      display_feats,
+      tooltipDiv,
+      renderTooltipDescription: this.renderTooltipDescription,
+      closeToolTip,
+      viewStart,
+      viewEnd,
+      source,
+      chr,
+    })
 
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    const renderTooltipDescription = this.renderTooltipDescription
-
-    const alreadyRendered = [] as string[] // hack fix for multiple transcript returns.
-    for (let i = 0; i < isoformData.length && row_count < MAX_ROWS; i++) {
-      const feature = isoformData[i]
-      let featureChildren = feature.children
-      if (featureChildren) {
-        const selected = feature.selected
-
-        // May want to remove this and add an external sort function
-        // outside of the render method to put certain features on top.
-        featureChildren = featureChildren.sort((a, b) =>
-          a.name.localeCompare(b.name),
-        )
-
-        // For each isoform..
-        let warningRendered = false
-        featureChildren.forEach(featureChild => {
-          if (
-            !(
-              isoformFilter.includes(featureChild.id) ||
-              isoformFilter.includes(featureChild.name)
-            ) &&
-            isoformFilter.length !== 0
-          ) {
-            return
-          }
-
-          if (alreadyRendered.includes(featureChild.id)) {
-            return
-          } else {
-            alreadyRendered.push(featureChild.id)
-          }
-          //
-          const featureType = featureChild.type
-
-          if (display_feats.includes(featureType)) {
-            // function to assign row based on available space.
-            // *** DANGER EDGE CASE ***/
-            let current_row = checkSpace(
-              used_space,
-              x(featureChild.fmin),
-              x(featureChild.fmax),
-            )
-            if (row_count < MAX_ROWS) {
-              // An isoform container
-              let text_string = ''
-              let text_label
-              let addingGeneLabel = false
-              const featName = feature.name
-              if (!Object.keys(geneList).includes(featName)) {
-                heightBuffer += GENE_LABEL_HEIGHT
-                addingGeneLabel = true
-                geneList[featName] = 'Green'
-              }
-
-              const isoform = track
-                .append('g')
-                .attr('class', 'isoform')
-                .attr(
-                  'transform',
-                  `translate(0,${row_count * ISOFORM_HEIGHT + 10 + heightBuffer})`,
-                )
-              if (addingGeneLabel) {
-                text_string = featName
-                text_label = isoform
-                  .append('text')
-                  .attr('class', 'geneLabel')
-                  .attr('fill', selected ? 'sandybrown' : 'black')
-                  .attr('height', ISOFORM_TITLE_HEIGHT)
-                  .attr(
-                    'transform',
-                    `translate(${x(featureChild.fmin)},-${GENE_LABEL_HEIGHT})`,
-                  )
-                  .text(text_string)
-                  .on('click', () => {
-                    renderTooltipDescription(
-                      tooltipDiv,
-                      renderTrackDescription(feature),
-                      closeToolTip,
-                    )
-                  })
-                  .datum({
-                    fmin: featureChild.fmin,
-                  })
-              }
-
-              isoform
-                .append('polygon')
-                .datum(() => ({
-                  fmin: featureChild.fmin,
-                  fmax: featureChild.fmax,
-                  strand: feature.strand,
-                }))
-                .attr('class', 'transArrow')
-                .attr('points', ARROW_POINTS)
-                .attr('transform', d =>
-                  feature.strand > 0
-                    ? `translate(${x(d.fmax)},0)`
-                    : `translate(${x(d.fmin)},${ARROW_HEIGHT}) rotate(180)`,
-                )
-                .on('click', () => {
-                  renderTooltipDescription(
-                    tooltipDiv,
-                    renderTrackDescription(featureChild),
-                    closeToolTip,
-                  )
-                })
-
-              isoform
-                .append('rect')
-                .attr('class', 'transcriptBackbone')
-                .attr('y', 10 + ISOFORM_TITLE_HEIGHT)
-                .attr('height', TRANSCRIPT_BACKBONE_HEIGHT)
-                .attr('transform', `translate(${x(featureChild.fmin)},0)`)
-                .attr('width', x(featureChild.fmax) - x(featureChild.fmin))
-                .on('click', () => {
-                  renderTooltipDescription(
-                    tooltipDiv,
-                    renderTrackDescription(featureChild),
-                    closeToolTip,
-                  )
-                })
-                .datum({
-                  fmin: featureChild.fmin,
-                  fmax: featureChild.fmax,
-                })
-
-              text_string = featureChild.name
-              text_label = isoform
-                .append('text')
-                .attr('class', 'transcriptLabel')
-                .attr('fill', selected ? 'sandybrown' : 'gray')
-                .attr('opacity', selected ? 1 : 0.5)
-                .attr('height', ISOFORM_TITLE_HEIGHT)
-                .attr('transform', `translate(${x(featureChild.fmin)},0)`)
-                .text(text_string)
-                .on('click', () => {
-                  renderTooltipDescription(
-                    tooltipDiv,
-                    renderTrackDescription(featureChild),
-                    closeToolTip,
-                  )
-                })
-                .datum({
-                  fmin: featureChild.fmin,
-                })
-
-              // Now that the label has been created we can calculate the space
-              // that this new element is taking up making sure to add in the
-              // width of the box.
-              // TODO: this is just an estimate of the length
-              let text_width = text_string.length * 2
-
-              // not some instances (as in reactjs?) the bounding box isn't
-              // available, so we have to guess
-              try {
-                text_width = text_label.node()?.getBBox().width ?? 0
-              } catch (e) {
-                // console.error('Not yet rendered',e)
-              }
-              // First check to see if label goes past the end
-              if (text_width + x(featureChild.fmin) > width) {
-                // console.error(featureChild.name + " goes over the edge");
-              }
-              const feat_end =
-                text_width > x(featureChild.fmax) - x(featureChild.fmin)
-                  ? x(featureChild.fmin) + text_width
-                  : x(featureChild.fmax)
-
-              // This is probably not the most efficient way to do this. Making
-              // an 2d array... each row is the first array (no zer0) next
-              // level is each element taking up space. Also using colons
-              // as spacers seems very perl... maybe change that?
-              // *** DANGER EDGE CASE ***/
-              if (used_space[current_row]) {
-                const temp = used_space[current_row]
-                temp.push(`${x(featureChild.fmin)}:${feat_end}`)
-                used_space[current_row] = temp
-              } else {
-                used_space[current_row] = [
-                  `${x(featureChild.fmin)}:${feat_end}`,
-                ]
-              }
-
-              // Now check on bounds since this feature is displayed
-              // The true end of display is converted to bp.
-              if (fmin_display < 0 || fmin_display > featureChild.fmin) {
-                fmin_display = featureChild.fmin
-              }
-              if (fmax_display < 0 || fmax_display < featureChild.fmax) {
-                fmax_display = featureChild.fmax
-              }
-
-              // have to sort this so we draw the exons BEFORE the CDS
-              if (featureChild.children) {
-                featureChild.children = featureChild.children.sort((a, b) => {
-                  const sortAValue = sortWeight[a.type]
-                  const sortBValue = sortWeight[b.type]
-
-                  if (
-                    typeof sortAValue === 'number' &&
-                    typeof sortBValue === 'number'
-                  ) {
-                    return sortAValue - sortBValue
-                  }
-                  if (
-                    typeof sortAValue === 'number' &&
-                    typeof sortBValue !== 'number'
-                  ) {
-                    return -1
-                  }
-                  if (
-                    typeof sortAValue !== 'number' &&
-                    typeof sortBValue === 'number'
-                  ) {
-                    return 1
-                  }
-                  // NOTE: type not found and weighted
-                  return a.type.localeCompare(b.type)
-                })
-
-                featureChild.children.forEach(innerChild => {
-                  const innerType = innerChild.type
-
-                  if (exon_feats.includes(innerType)) {
-                    isoform
-                      .append('rect')
-                      .attr('class', 'exon')
-                      .attr('x', x(innerChild.fmin))
-                      .attr(
-                        'transform',
-                        `translate(0,${EXON_HEIGHT - TRANSCRIPT_BACKBONE_HEIGHT})`,
-                      )
-                      .attr('height', EXON_HEIGHT)
-                      .attr('z-index', 10)
-                      .attr('width', x(innerChild.fmax) - x(innerChild.fmin))
-                      .on('click', () => {
-                        renderTooltipDescription(
-                          tooltipDiv,
-                          renderTrackDescription(featureChild),
-                          closeToolTip,
-                        )
-                      })
-                      .datum({ fmin: innerChild.fmin, fmax: innerChild.fmax })
-                  } else if (CDS_feats.includes(innerType)) {
-                    isoform
-                      .append('rect')
-                      .attr('class', 'CDS')
-                      .attr('x', x(innerChild.fmin))
-                      .attr(
-                        'transform',
-                        `translate(0,${CDS_HEIGHT - TRANSCRIPT_BACKBONE_HEIGHT})`,
-                      )
-                      .attr('z-index', 20)
-                      .attr('height', CDS_HEIGHT)
-                      .attr('width', x(innerChild.fmax) - x(innerChild.fmin))
-                      .on('click', () => {
-                        renderTooltipDescription(
-                          tooltipDiv,
-                          renderTrackDescription(featureChild),
-                          closeToolTip,
-                        )
-                      })
-                      .datum({ fmin: innerChild.fmin, fmax: innerChild.fmax })
-                  } else if (UTR_feats.includes(innerType)) {
-                    isoform
-                      .append('rect')
-                      .attr('class', 'UTR')
-                      .attr('x', x(innerChild.fmin))
-                      .attr(
-                        'transform',
-                        `translate(0,${UTR_HEIGHT - TRANSCRIPT_BACKBONE_HEIGHT})`,
-                      )
-                      .attr('z-index', 20)
-                      .attr('height', UTR_HEIGHT)
-                      .attr('width', x(innerChild.fmax) - x(innerChild.fmin))
-                      .on('click', () => {
-                        renderTooltipDescription(
-                          tooltipDiv,
-                          renderTrackDescription(featureChild),
-                          closeToolTip,
-                        )
-                      })
-                      .datum({ fmin: innerChild.fmin, fmax: innerChild.fmax })
-                  }
-                })
-              }
-              row_count += 1
-            }
-            if (row_count === MAX_ROWS && !warningRendered) {
-              // *** DANGER EDGE CASE ***/
-              const link = getJBrowseLink(source, chr, viewStart, viewEnd)
-              ++current_row
-              warningRendered = true
-              track
-                .append('a')
-                .attr('class', 'transcriptLabel')
-                .attr('xlink:show', 'new')
-                .append('text')
-                .attr('x', 10)
-                .attr('y', 10)
-                .attr(
-                  'transform',
-                  `translate(0,${row_count * ISOFORM_HEIGHT + 20 + heightBuffer})`,
-                )
-                .attr('fill', 'red')
-                .attr('opacity', 1)
-                .attr('height', ISOFORM_TITLE_HEIGHT)
-                .html(link)
-            }
-          }
-        })
-      }
-    }
     if (initialHighlight) {
       setHighlights(initialHighlight, viewer)
     }
 
-    if (row_count === 0) {
-      track
-        .append('text')
-        .attr('x', 30)
-        .attr('y', ISOFORM_TITLE_HEIGHT + 10)
-        .attr('fill', 'orange')
-        .attr('opacity', 0.6)
-        .text(
-          'Overview of non-coding genome features unavailable at this time.',
-        )
-    }
     // we return the appropriate height function
-    return row_count * ISOFORM_HEIGHT + heightBuffer + VARIANT_TRACK_HEIGHT
+    return returnedHeight + VARIANT_TRACK_HEIGHT
   }
 
-  filterVariantData(variantData: VariantFeature[], variantFilter: string[]) {
-    if (variantFilter.length === 0) {
-      return variantData
-    }
-    return variantData.filter(v => {
-      let returnVal = false
-      try {
-        if (
-          variantFilter.includes(v.name) ||
-          (v.allele_symbols?.values &&
-            variantFilter.includes(
-              v.allele_symbols.values[0].replace(/"/g, ''),
-            )) ||
-          (v.symbol?.values &&
-            variantFilter.includes(v.symbol.values[0].replace(/"/g, ''))) ||
-          (v.symbol_text?.values &&
-            variantFilter.includes(v.symbol_text.values[0].replace(/"/g, '')))
-        ) {
-          returnVal = true
-        }
-        const ids =
-          v.allele_ids?.values[0].replace(/"|\[|\]| /g, '').split(',') ?? []
-        ids.forEach(id => {
-          if (variantFilter.includes(id)) {
-            returnVal = true
-          }
-        })
-      } catch (e) {
-        console.error(
-          'error processing filter with so returning anyway',
-          variantFilter,
-          v,
-          e,
-        )
-        returnVal = true
-      }
-      return returnVal
-    })
-  }
+
 
   renderTooltipDescription(
     tooltipDiv: Selection<HTMLDivElement, unknown, HTMLElement, undefined>,
