+++ src/tracks/IsoformEmbeddedVariantTrack.ts
@@ -8,6 +8,7 @@ import {
 } from '../RenderFunctions'
 import { renderTrackDescription } from '../services/TrackService'
 import {
+  filterVariantData,
   generateVariantDataBinsAndDataSets,
   getColorsForConsequences,
   getVariantAlleles,
@@ -16,21 +17,15 @@ import {
   renderVariantDescriptions,
 } from '../services/VariantService'
 
+import BaseTrack from './BaseTrack'
+
 import type { VariantFeature } from '../services/VariantService'
 import type { SimpleFeatureSerialized } from '../services/types'
 
-export default class IsoformEmbeddedVariantTrack {
+export default class IsoformEmbeddedVariantTrack extends BaseTrack {
   private trackData: SimpleFeatureSerialized[]
   private variantData: VariantFeature[]
-  private viewer: d3.Selection<
-    SVGGElement,
-    unknown,
-    HTMLElement | null,
-    undefined
-  >
-  private width: number
   private variantFilter: string[]
-  private height: number
   private transcriptTypes: string[]
   private variantTypes: string[]
   private showVariantLabel: boolean
@@ -59,13 +54,11 @@ export default class IsoformEmbeddedVariantTrack {
     variantData?: VariantFeature[]
     trackData?: SimpleFeatureSerialized[]
   }) {
+    super({ viewer, width, height })
     this.trackData = trackData ?? []
     this.variantData = variantData ?? []
-    this.viewer = viewer
-    this.width = width
     this.variantFilter = variantFilter
     this.initialHighlight = initialHighlight
-    this.height = height
     this.transcriptTypes = transcriptTypes
     this.variantTypes = variantTypes
     this.showVariantLabel = showVariantLabel ?? true
@@ -76,7 +69,7 @@ export default class IsoformEmbeddedVariantTrack {
     const trackData = this.trackData
     let isoformData = trackData
 
-    const variantData = this.filterVariantData(
+    const variantData = filterVariantData(
       variantDataPre,
       this.variantFilter,
     )
@@ -689,94 +682,7 @@ export default class IsoformEmbeddedVariantTrack {
     // we return the appropriate height function
     return row_count * ISOFORM_HEIGHT + heightBuffer
   }
-  private filterVariantData(
-    variantData: VariantFeature[],
-    variantFilter: string[],
-  ): VariantFeature[] {
-    if (variantFilter.length === 0) {
-      return variantData
-    }
 
-    // Convert filter array to Set for O(1) lookups
-    const filterSet = new Set(variantFilter)
-
-    const filteredResults = variantData.filter(v => {
-      let returnVal = false
-      try {
-        // Check name match
-        if (filterSet.has(v.name)) {
-          returnVal = true
-        }
-
-        // Check allele_symbols match
-        if (v.allele_symbols?.values) {
-          const cleanedSymbol = v.allele_symbols.values[0].replace(
-            /"|\\[|\\]| /g,
-            '',
-          )
-          if (filterSet.has(cleanedSymbol)) {
-            returnVal = true
-          }
-        }
-
-        // Check symbol match
-        if (v.symbol?.values) {
-          const cleanedSymbol = v.symbol.values[0].replace(/"|\\[|\\]| /g, '')
-          if (filterSet.has(cleanedSymbol)) {
-            returnVal = true
-          }
-        }
-
-        // Check symbol_text match
-        if (v.symbol_text?.values) {
-          const cleanedSymbolText = v.symbol_text.values[0].replace(
-            /"|\\[|\\]| /g,
-            '',
-          )
-          if (filterSet.has(cleanedSymbolText)) {
-            returnVal = true
-          }
-        }
-
-        // Handle allele_ids with JSON parsing support
-        // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-        const rawValue = v.allele_ids?.values?.[0]
-
-        if (rawValue) {
-          let ids: string[] = []
-
-          // Check if it's a JSON stringified array
-          if (rawValue.startsWith('[') && rawValue.endsWith(']')) {
-            try {
-              const parsed: unknown = JSON.parse(rawValue)
-              ids = (Array.isArray(parsed) ? parsed : [parsed]).map(String)
-            } catch (e) {
-              // Fallback to original parsing
-              ids = rawValue.replace(/"|\\[|\\]| /g, '').split(',')
-            }
-          } else {
-            // Original parsing logic
-            ids = rawValue.replace(/"|\\[|\\]| /g, '').split(',')
-          }
-
-          // Use Set.has() for O(1) lookup
-          for (const id of ids) {
-            if (filterSet.has(id)) {
-              returnVal = true
-              break
-            }
-          }
-        }
-      } catch (e) {
-        // On error, include the variant
-        returnVal = true
-      }
-
-      return returnVal
-    })
-
-    return filteredResults
-  }
 
   private renderTooltipDescription(
     tooltipDiv: d3.Selection<
